# openapi.yaml
openapi: 3.0.0
info:
  title: educultura API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3018/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /material-multimedia:
    get:
      tags:
        - MaterialMultimedia
      summary: Listar todos los MaterialMultimedias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de MaterialMultimedias
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - MaterialMultimedia
      summary: Crear MaterialMultimedia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: MaterialMultimedia creado
  /material-multimedia/{id}:
    get:
      tags:
        - MaterialMultimedia
      summary: Obtener MaterialMultimedia por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MaterialMultimedia encontrado
        '404':
          description: MaterialMultimedia no encontrado
    patch:
      tags:
        - MaterialMultimedia
      summary: Actualizar MaterialMultimedia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: MaterialMultimedia actualizado
    delete:
      tags:
        - MaterialMultimedia
      summary: Eliminar MaterialMultimedia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MaterialMultimedia eliminado
  /material-multimedia/{id}/upload-image:
    post:
      tags:
        - MaterialMultimedia
      summary: Subir imagen a MaterialMultimedia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida